requirements.txt
boto3==1.34.19
PyYAML==6.0.1